Cast constant to avoid integer overflow
Since VmemTracker_GetChunkSizeInBits() returns int32, the compiler
will use 32-bit integer arithmetic for the shift, even though the
resulting variable is 64-bit wide. Cast the constant to int64 to
ensure we don't risk a 32-bit overflow where we can store it. This
matches other uses of VmemTracker_GetChunkSizeInBits() where we
want a 64-bit result.

Reported-by: Coverity static analysis.

+++ src/backend/utils/resgroup/resgroup.c
@@ -1211,7 +1211,7 @@ ResourceGroupGetQueryMemoryLimit(void)
 	if (bypassedGroup)
 	{
 		int64		bytesInMB = 1 << BITS_IN_MB;
-		int64		bytesInChunk = 1 << VmemTracker_GetChunkSizeInBits();
+		int64		bytesInChunk = (int64) 1 << VmemTracker_GetChunkSizeInBits();
 
 		/*
 		 * In bypass mode there is a hard memory limit of
